### 2024-04-20 #6649 @arfon
@hanecakr – could you link me to the rOpenSci review thread please?

Also, please shorten your paper. JOSS submissions should be ~1000 words.
