### 2024-04-20 #6649 @hanecakr
Dear @arfon, thank you for accepting to act as an editor for my submission.

The rOpenSci review can be found here: https://github.com/ropensci/software-review/issues/618

I'll try to shorten the paper in the next few days and let you know when it's ready.
