+++ pkgdown.yml
@@ -1,9 +1,9 @@
 pandoc: 3.1.11
-pkgdown: 2.0.8
+pkgdown: 2.0.9
 pkgdown_sha: ~
 articles:
   getting_started: getting_started.html
-last_built: 2024-04-16T20:00Z
+last_built: 2024-04-23T06:04Z
 urls:
   reference: https://hanecakr.github.io/fellingdater/reference
   article: https://hanecakr.github.io/fellingdater/articles
