### 2024-04-24 bff1c93
shortened version of paper.md
