+++ inst/paper/paper.bib
@@ -320,19 +320,6 @@
 	doi = {https://doi.org/10.1111/1475-4754.00048},
 }
 
-@article{haneca2014,
-	title = {Tree-Rings, Timbers and Trees: a dendrochronological survey of the 14th-century cog, Doel 1},
-	author = {Haneca, K. and Daly, A.},
-	year = {2014},
-	date = {2014},
-	journal = {International Journal of Nautical Archaeology},
-	pages = {87--102},
-	volume = {43},
-	number = {1},
-	doi = {10.1111/1095-9270.12037},
-	url = {http://doi.wiley.com/10.1111/1095-9270.12037}
-}
-
 @article{haneca2020,
 	title = {The ups and downs of the building trade in a medieval city: Tree-ring data as proxies for economic, social and demographic dynamics in Bruges (c. 1200{\textendash}1500)},
 	author = {Haneca, K. and Debonne, V. and Hoffsummer, P.},
