+++ inst/paper/paper.md
@@ -1,25 +1,23 @@
 ---
-title: "_fellingdater_: a toolkit to estimate, report and combine felling dates derived from historical tree-ring series."
+title: "_fellingdater_: a toolkit to estimate, report and combine felling dates derived
+  from historical tree-ring series."
 authors:
-  - name: Kristof Haneca
-    orcid: 0000-0002-7719-8305
-    affiliation: 1
+- name: Kristof Haneca
+  orcid: "0000-0002-7719-8305"
+  affiliation: 1
 affiliations:
-  - name: Flanders Heritage Agency, Belgium
-    index: 1
-date: 2024-04-21
-output: 
-  pdf_document: default
-  html_document:
-  df_print: paged
-tags: 
-  - R
-  - tree-ring analysis
-  - dendrochronology
-  - sapwood
-  - dendroarchaeology
-editor_options: 
-  markdown: 
+- name: Flanders Heritage Agency, Belgium
+  index: 1
+date: "2024-04-24"
+output: pdf_document
+tags:
+- R
+- tree-ring analysis
+- dendrochronology
+- sapwood
+- dendroarchaeology
+editor_options:
+  markdown:
     wrap: 72
 bibliography: paper.bib
 ---
@@ -27,82 +25,50 @@ bibliography: paper.bib
 # Summary
 
 Tree-ring dating, or dendrochronology, allows to assign calendar-year
-dates to growth rings that are observable on a cross-section of a stem
-or a piece of timber. It involves measuring the width of each growth
-ring and comparing the measured ring-width pattern to absolutely dated
-reference chronologies. Once a tree-ring series is securely anchored to
-a calendar year time-scale, the end date of the outermost ring can be
-used to determine or estimate the year of death of the parent tree
-(i.e., the felling of the tree).
+dates to growth rings that can be observed on an old piece of timber.
+Once a tree-ring series is securely anchored to a calendar year
+time-scale, the end date of the outermost ring can be used to estimate
+the year when the tree was felled.
 
-The `fellingdater` package aims to offer a suite of functions that can
-assist dendrochronologists to infer, combine and report felling date
-estimates from dated tree-ring series of (pre-)historical timbers, based
-on the presence of partially preserved sapwood or waney edge ([Fig.
-1](#fig-cross-section)).
+The `fellingdater` package offers a suite of functions that can assist
+dendrochronologists to infer, combine and report felling date estimates
+from dated tree-ring series, based on the presence of partially
+preserved sapwood or waney edge ([Fig. 1](#fig-cross-section)).
 
 ![A cross-section of a historical timber from a medieval roof
-construction. The heartwood and sapwood areas are marked, as well as all
-ring boundaries. The `fellingdater` package offers a workflow to
-estimate the number of missing sapwood rings -- that have been
-trimmed-of, powdered by wood-boring insects, or deteriorated by
-wood-decaying fungi -- between the last measured ring and the cambial
-layer.](cross-section.png){#fig-cross-section}
+construction. All ring boundaries are marked, as well as the heartwood
+and the partially preserved
+sapwood.](cross-section.png){#fig-cross-section}
 
 # Background
 
 Dendrochronology is the most precise chronometric dating technique for
 (pre-)historical wooden constructions and objects [@baillie1995]. It
-involves recording the ring-width pattern on a cross-section of the
-wooden element under study and matching this pattern to absolutely dated
-reference chronologies, which allows the recorded tree-ring pattern to
-be anchored to an absolute time scale. From a dated tree-ring pattern it
-is known in which growing season each growth ring has been laid down by
-the parent tree. In archaeological, architectural of art-historical
-studies, the primary objective of a dendrochronological survey is to
-ascertain an accurate estimate of the **felling date** (or dying-off) of
-the parent tree from which the timber originates [@haneca2009;
[…]
