### 2024-04-24 #6649 @hanecakr
Dear @afron, I've shortened the paper. By including an example of use, it is still longer than 1000 words but hope this section can be included as it illustrates the main functions of the package.
