### 2024-04-26 722b6a9
Update _pkgdown.yml
Sorry, made a mistake in my own repo after noticing the same with yours...
