+++ _pkgdown.yml
@@ -1,4 +1,4 @@
-url: https://ropensci.github.io/fellingdater/
+url: https://https://docs.ropensci.org/fellingdater/
 template:
   bootstrap: 5
 reference:
