### 2024-05-02 #6649 @arfon
@hanecakr - thanks for suggesting additional reviewers but we don't need new ones (or any further review) as the rOpenSci review meets our needs @ JOSS
