### 2024-05-02 #6716 @editorialbot
[REVIEW]: fellingdater: a toolkit to estimate, report and combine felling dates derived from historical tree-ring series.
**Submitting author:** @hanecakr (<a href="http://orcid.org/0000-0002-7719-8305">Kristof Haneca</a>)
**Repository:** https://github.com/ropensci/fellingdater
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.1
