+++ .github/workflows/test-coverage.yaml
@@ -1,10 +1,4 @@
-steps:
-- uses: actions/checkout@main
-- uses: codecov/codecov-action@v4
-  with:
-    fail_ci_if_error: true # optional (default = false)
-    files: ./coverage1.xml,./coverage2.xml # optional
-    flags: unittests # optional
-    name: codecov-umbrella # optional
-    token: ${{ secrets.CODECOV_TOKEN }} # required
-    verbose: true # optional (default = false)
+- name: Upload coverage reports to Codecov
+    uses: codecov/codecov-action@v4.0.1
+    with:
+      token: ${{ secrets.CODECOV_TOKEN }}
