### 2024-05-04 fd45280
remove codecoverage badge from README
