+++ inst/paper/paper.bib
@@ -85,7 +85,8 @@
 	journal = {Journal of Forest Science},
 	pages = {141--146},
 	volume = {52},
-	number = {3}
+	number = {3},
+	doi = {10.17221/4496-JFS}
 }
 
 @article{pilcher1987,
@@ -149,7 +150,6 @@
 	volume = {30},
 	number = {1},
 	doi = {10.1016/j.dendro.2011.08.001},
-	url = {http://linkinghub.elsevier.com/retrieve/pii/S1125786511000750},
 	langid = {en}
 }
 
@@ -170,7 +170,8 @@
 	date = {1981},
 	journal = {Journal of Archaeological Science},
 	pages = {381--390},
-	volume = {8}
+	volume = {8},
+	doi = {10.1016/0305-4403(81)90037-6}
 }
 
 @article{jevsenak2019,
@@ -183,7 +184,6 @@
 	pages = {64--70},
 	volume = {54},
 	doi = {10.1016/j.dendro.2019.02.006},
-	url = {https://www.sciencedirect.com/science/article/pii/S1125786518301966},
 	langid = {en}
 }
 
@@ -208,7 +208,6 @@
 	volume = {31},
 	number = {3},
 	doi = {10.1016/j.dendro.2013.03.001},
-	url = {http://linkinghub.elsevier.com/retrieve/pii/S1125786513000246},
 	langid = {en}
 }
 
@@ -221,7 +220,7 @@
 	pages = {137--143},
 	volume = {44},
 	number = {1},
-	doi = {https://doi.org/10.1111/1475-4754.00048},
+	doi = {10.1111/1475-4754.00048},
 }
 
 @article{miles2006,
@@ -244,7 +243,7 @@
 	pages = {91{\textendash}106},
 	volume = {39},
 	number = {1},
-	doi = {https://doi.org/10.1179/174962908X365082}
+	doi = {10.1179/174962908X365082}
 }
 
 @book{rcoreteam2022,
@@ -317,7 +316,7 @@
 	pages = {137--143},
 	volume = {44},
 	number = {1},
-	doi = {https://doi.org/10.1111/1475-4754.00048},
+	doi = {10.1111/1475-4754.00048},
 }
 
 @article{haneca2020,
@@ -365,8 +364,7 @@
 	journal = {Dendrochronologia},
 	pages = {29--30},
 	volume = {34},
-	doi = {10.1016/j.dendro.2015.03.003},
-	url = {http://www.sciencedirect.com/science/article/pii/S1125786515000193}
+	doi = {10.1016/j.dendro.2015.03.003}
 }
 
 @article{eckstein1969,
@@ -377,7 +375,7 @@
 	journal = {Forstwissenschaftliches Centralblatt},
 	pages = {230--250},
 	volume = {88},
-	doi = {https://doi.org/10.1007/BF02741777}
+	doi = {10.1007/BF02741777}
 }
 
 @article{visser2009,
@@ -399,7 +397,8 @@
 	journal = {Holz als Roh und Werkstoff},
 	pages = {263--268},
 	volume = {6},
-	number = {10-12}
+	number = {10-12},
+	doi = {10.1007/BF02603303}
 }
 
 @article{visser2021,
