### 2024-05-04 #6716 @hanecakr
@arfon the doi's that were missing or needed some tweeking were addressed in this commit: https://github.com/ropensci/fellingdater/commit/759b253ac35164da35e4219ef4107f90da0ef856
