### 2024-05-12 4c3d8bc
fixing typo's in paper.md
