+++ inst/paper/paper.md
@@ -47,13 +47,13 @@ Dendrochronology is the most precise chronometric dating technique for
 (pre-)historical wooden constructions and objects [@baillie1995]. It
 involves recording the ring-width pattern on a cross-section of an
 ancient wooden element and matching this pattern to absolutely dated
-reference chronologies. This allows to anchor the recorded tree-ring
+reference chronologies. This allows anchoring the recorded tree-ring
 pattern to an absolute time scale. In archaeological, architectural or
 art-historical studies, the primary objective of a dendrochronological
 survey is to ascertain an accurate estimate of the **felling date** (or
 dying-off) of the parent tree from which the timber originates
-[@haneca2009; [@dominguez-delmas_2020]; @tegel2022]. This felling date
-is the closest related and datable event to the creation of the wooden
+[@haneca2009; @dominguez-delmas_2020; @tegel2022]. This felling date is
+the closest related and datable event to the creation of the wooden
 object or construction.
 
 The exact felling date can be inferred from the most recently formed
@@ -74,39 +74,40 @@ based on the presence of (partially) preserved sapwood or waney edge.
 # Statement of need
 
 Many descriptive statistics and statistical models have been published
-to establish accurate estimates of the expected number of sapwood ring
+to establish accurate estimates of the expected number of sapwood rings
 [@edvardsson2022; @bleicher2020; @rybnicek2006; @pilcher1987;
 @hollstein1965; @hollstein1980; @wazny1990; @miles1997; @sohar2012;
 @bräthen1982; @haneca2009; @hughes1981; @jevsenak2019; @hillam1987;
-@gjerdrum2013; @shindo2024]. These models often rely on
-log-transformation of the original counts of sapwood numbers from living
-and historical timbers, or use regression models that include additional
-variables such as mean ring width, the cambial age of the tree or a
-combination of both. A standardized methodology is hampered by this
-variety in methodology and reporting.
+@gjerdrum2013; @shindo2024]. These models are based on counts of sapwood
+rings from living and historical timbers and often rely on
+log-transformation of the original data, or use regression models
+including additional variables such as mean ring width, the cambial age
+of the tree or a combination of both. A standardized methodology for
+reporting felling dates is therefore hampered by this variety in
+statistical approaches.
 
 A Bayesian method to improve the procedures to model sapwood data,
 compute lower and upper limits for the felling date based upon the
 selected sapwood model and a given credible interval have been
-introduced by @millard2002a. This procedure was then further refined by
+introduced by @millard2002a. This procedure was further refined by
 @miles2006, and critically reviewed with real-life examples by
 @tyers2008. This workflow has been incorporated in
 [OxCal](https://c14.arch.ox.ac.uk/oxcalhelp/Sapwood.html), the routine
 software for radiocarbon dating and modelling [@bronkramsey2009].
 Tree-ring analysis, on the other hand, relies on a growing set of
 R-packages, with the '*Dendrochronology Program Library in R*'
-[@bunn2008; @bunn2010; @bunn2022], at it's core (
-[opendendro.org](https://opendendro.org/); [@bunn2022a]). Yet, the
+[@bunn2008; @bunn2010; @bunn2022], at it's core
+[https://opendendro.org/, @bunn2022a]. Yet, the
 reporting of felling dates is currently not a standardized procedure
 incorporated in an R-package.
 
-The `fellingdater`package allows to fully document the methodology to
+The `fellingdater` package allows to fully document the methodology to
 establish a felling date -- for a single timber or a group of timbers --
 making the whole procedure reproducible and assists in building
-standardized workflows when applied to large datasets (e.g.
[…]
