### 2024-05-13 d09c810
New vignette (paper) "The `fellingdater` workflow"
