+++ vignettes/articles/the_workflow.Rmd
@@ -1,0 +1,310 @@
+---
+title: "The `fellingdater` workflow"
+bibliography: workflow_vignette.bib
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+```{r setup, echo=FALSE}
+library(fellingdater)
+```
+
+> **This vignette is a lengthy version of a paper submitted to '*The Journal of Open Source Software*' (JOSS).**
+
+# Summary
+
+Tree-ring dating, or dendrochronology, allows to assign calendar-year dates to growth rings that can be observed on a cross-section of a stem or a piece of timber. It involves measuring the width of each growth ring and comparing the measured ring-width pattern to absolutely dated reference chronologies. Once a tree-ring series is securely anchored to a calendar year time-scale, the end date of the outermost ring can be used to determine or estimate the year of death of the parent tree (i.e., the felling of the tree).
+
+The `fellingdater` package offers a suite of functions that can assist dendrochronologists to infer, combine and report felling date estimates from dated tree-ring series of (pre-)historical timbers, based on the presence of partially preserved sapwood or waney edge ([Fig. 1](#fig-cross-section)).
+
+![Fig. 1: A cross-section of a historical timber from a medieval roof construction. All ring boundaries are marked, as well as the heartwood and the partially preserved sapwood.](cross-section.png){#fig-cross-section}
+
+# Background
+
+Dendrochronology is the most precise chronometric dating technique for (pre-)historical wooden constructions and objects [@baillie1995]. It involves recording the ring-width pattern on a cross-section of an ancient wooden element and matching this pattern to absolutely dated reference chronologies, which allows anchoring the recorded tree-ring pattern to an absolute time scale. From a dated tree-ring pattern it is known in which growing season each growth ring has been laid down by the parent tree. In archaeological, architectural of art-historical studies, the primary objective of a dendrochronological survey is to ascertain an accurate estimate of the **felling date** (or dying-off) of the parent tree from which the timber originates [@haneca2009; @dominguez-delmas_2020; @tegel2022]. This felling date is the closest related and datable event to the creation of the wooden object or construction. These precisely dated events serves as the basis for narratives on various aspects, such as timber selection, craftsmanship, workshop practices, trade, provenance, and historical forest management [@dominguez-delmas-etal_2023].
+
[…]
