+++ vignettes/articles/workflow_vignette.bib
@@ -1,0 +1,541 @@
+
+@article{haneca2012,
+	title = {Precise tree-ring dating of building activities despite the absence of bark: A case-study on medieval church roofs in Damme, Belgium},
+	author = {Haneca, K. and Debonne, V.},
+	year = {2012},
+	date = {2012},
+	journal = {Dendrochronologia},
+	pages = {23--34},
+	volume = {30},
+	number = {1},
+	doi = {10.1016/j.dendro.2011.06.002},
+}
+
+@book{baillie1995,
+	title = {A slice through time. Dendrochronology and precision dating},
+	author = {Baillie, M.G.L.},
+	year = {1995},
+	date = {1995},
+	publisher = {B.T. Batsford Ltd},
+	address = {London}
+}
+
+@article{haneca2009,
+	title = {Oaks, tree-rings and wooden cultural heritage: a review of the main characteristics and applications of oak dendrochronology in Europe},
+	author = {Haneca, K. and Cufar, K. and Beeckman, H.},
+	year = {2009},
+	month = {01},
+	date = {2009-01},
+	journal = {Journal of Archaeological Science},
+	pages = {1--11},
+	volume = {36},
+	number = {1},
+	doi = {10.1016/j.jas.2008.07.005},
+	note = {WOS:000260986300001},
+	langid = {English}
+}
+
+@article{tegel2022,
+	title = {Dendroarchaeology in Europe},
+	author = {Tegel, W. and Muigg, B. and Skiadaresis, G. and Vanmoerkerke, J. and Seim, A.},
+	year = {2022},
+	month = {02},
+	date = {2022-02-16},
+	journal = {Frontiers in Ecology and Evolution},
+	pages = {823622},
+	volume = {10},
+	doi = {10.3389/fevo.2022.823622},
+	url = {https://www.frontiersin.org/articles/10.3389/fevo.2022.823622/full},
+	langid = {en}
+}
+
+@article{edvardsson2022,
+	title = {A statistical model for the prediction of the number of sapwood rings in Scots pine (Pinus sylvestris L.)},
+	author = {Edvardsson, J. and {Rögnvaldsson}, K. and {Helgadóttir}, E.D. and Linderson, H. and Hrafnkelsson, B.},
+	year = {2022},
+	month = {08},
+	date = {2022},
+	journal = {Dendrochronologia},
+	pages = {125963},
+	volume = {74},
+	doi = {10.1016/j.dendro.2022.125963},
+	url = {https://linkinghub.elsevier.com/retrieve/pii/S1125786522000431},
+	langid = {en}
+}
+
+@article{bleicher2020,
+	title = {The Zurich method for sapwood estimation},
+	author = {Bleicher, N. and Walder, F. and Gut, U. and Bolliger, M.},
+	year = {2020},
+	month = {12},
+	date = {2020-12-01},
+	journal = {Dendrochronologia},
+	pages = {125776},
+	volume = {64},
+	doi = {10.1016/j.dendro.2020.125776},
+	url = {http://www.sciencedirect.com/science/article/pii/S1125786520301156},
+	langid = {en}
+}
+
+@article{rybnicek2006,
+	title = {Determination of the number of sapwood annual rings in oak in the region of Southern Moravia},
+	author = {Rybnicek, M. and Vavrik, H. and Hubeny, R.},
+	year = {2006},
+	date = {2006},
+	journal = {Journal of Forest Science},
+	pages = {141--146},
+	volume = {52},
+	number = {3}
+}
+
+@article{pilcher1987,
[…]
