+++ _pkgdown.yml
@@ -1,4 +1,4 @@
-url: https://hanecakr.github.io/fellingdater/
+url: https://ropensci.github.io/fellingdater/
 template:
   bootstrap: 5
 reference:
