### 2024-05-13 3e19b9d
correct path to illustrations for new vignette
