+++ vignettes/articles/the_workflow.Rmd
@@ -1,5 +1,5 @@
 ---
-title: "The `fellingdater` workflow"
+title: "The ´fellingdater´ workflow"
 bibliography: workflow_vignette.bib
 ---
 
@@ -16,13 +16,17 @@ library(fellingdater)
 
 > **This vignette is a lengthy version of a paper submitted to '*The Journal of Open Source Software*' (JOSS).**
 
+\
+
 # Summary
 
 Tree-ring dating, or dendrochronology, allows to assign calendar-year dates to growth rings that can be observed on a cross-section of a stem or a piece of timber. It involves measuring the width of each growth ring and comparing the measured ring-width pattern to absolutely dated reference chronologies. Once a tree-ring series is securely anchored to a calendar year time-scale, the end date of the outermost ring can be used to determine or estimate the year of death of the parent tree (i.e., the felling of the tree).
 
 The `fellingdater` package offers a suite of functions that can assist dendrochronologists to infer, combine and report felling date estimates from dated tree-ring series of (pre-)historical timbers, based on the presence of partially preserved sapwood or waney edge ([Fig. 1](#fig-cross-section)).
 
-![Fig. 1: A cross-section of a historical timber from a medieval roof construction. All ring boundaries are marked, as well as the heartwood and the partially preserved sapwood.](cross-section.png){#fig-cross-section}
+![Fig. 1: A cross-section of a historical timber from a medieval roof construction. All ring boundaries are marked, as well as the heartwood and the partially preserved sapwood.](images/cross-section-01.png){#fig-cross-section}
+
+\
 
 # Background
 
@@ -34,6 +38,8 @@ To refine estimates of felling dates, since the early development of tree-ring d
 
 The `fellingdater` package aims to facilitate this process by providing functions to infer, combine and report felling date estimates from dated tree-ring series, based on the presence of (partially) preserved sapwood or waney edge.
 
+\
+
 # Statement of need
 
 Many descriptive statistics and statistical models have been published to establish accurate estimates of the expected number of sapwood ring [@edvardsson2022; @bleicher2020; @rybnicek2006; @pilcher1987; @hollstein1965; @hollstein1980; @wazny1990; @miles1997; @sohar2012; @bräthen1982; @haneca2009; @hughes1981; @jevsenak2019; @hillam1987; @gjerdrum2013; @shindo2024]. These models are based on counts of sapwood rings from living and historical timber and often rely on log-transformation of the original data, or use regression models including additional variables such as mean ring width, the cambial age of the tree or a combination of both. These statistical procedures report the expected minimal and maximal number of sapwood rings, usually within 95% a confidence interval, but have also been presented in a wide variety of ways and differ among laboratories and dendrochronologists. A standardized methodology for reporting felling dates is therefore hampered by this variety in statistical approaches. This variety in methodology and reporting comes even more to the surface when tree-ring dates of multiple elements from a single object, construction or building phase are combined into a single felling date for the whole ensemble. The goal of such a mutual interpretation of the individual felling dates is to refine the range of the felling date, but also to check or test whether these dated tree-ring series/wooden elements could indeed represent one single event (i.e. the felling of trees).
@@ -42,7 +48,9 @@ A Bayesian method to improve the procedures to model sapwood data, compute lower
 
[…]
