+++ pkgdown.yml
@@ -4,7 +4,7 @@ pkgdown_sha: ~
 articles:
   getting_started: getting_started.html
   the_workflow: the_workflow.html
-last_built: 2024-05-13T13:06Z
+last_built: 2024-05-13T14:02Z
 urls:
   reference: https://ropensci.github.io/fellingdater/reference
   article: https://ropensci.github.io/fellingdater/articles
