### 2024-05-13 #6716 @hanecakr
@arfon: I fixed some typo's (especially related to the bibliographic references) in paper.md
-> commit https://github.com/ropensci/fellingdater/commit/4c3d8bc7afc45c804ef9b059dc21331b577b9aa8
