### 2024-05-13 #6716 @hanecakr
`fellingdater` v1.0.1 has been released and archived to Zenodo: https://doi.org/10.5281/zenodo.11186275
