+++ NEWS.md
@@ -1,5 +1,6 @@
 # fellingdater 1.0.1
 
+
 fellingdater 1.0.1 (2024-05-13)
 ===============================
   * Paper.md updated for ‘The Journal of Open Source Software’ (JOSS).
