### 2024-05-14 #6716 @arfon
@hanecakr – looks like some of the internal references to figures aren't working. Can you take a look at these docs and update accordingly? https://joss.readthedocs.io/en/latest/paper.html#tables-and-figures
