### 2024-05-14 cc02eb2
fix figure reference in paper.md
