### 2024-05-18 #6716 @arfon
@hanecakr – I made a few small edits to the paper in https://github.com/ropensci/fellingdater/pull/16 – could you please review and merge?
