### 2024-05-22 0fb6da0
update CITATION - paper.md accepted for publication in Journal of Open Source Software
