### 2024-05-22 41ce14f
update of README following the publication in JOSS
