+++ NEWS.md
@@ -1,38 +1,34 @@
-# fellingdater 1.0.1
+# fellingdater 1.0.2
 
+# fellingdater 1.0.2 (2024-05-22)
 
-fellingdater 1.0.1 (2024-05-13)
-===============================
-  * Paper.md updated for ‘The Journal of Open Source Software’ (JOSS).
-  * New vignette (article): "The fellingdater workflow"
-  
+### PUBLICATION
 
-fellingdater 1.0.0 (2024-04-16)
-===============================
+-   Paper accepted for publication in 'The Journal of Open Source Software' (JOSS).
+-   doi: <https://doi.org/10.21105/joss.06716>
+-   CITATION and README updated accordingly
 
-  * Major release of `fellingdater` to reflect the fact that the package is now 
-  fully functional and has successfully passed the rOpenSci review process.
-  * review history: https://github.com/ropensci/software-review/issues/618
-  
+# fellingdater 1.0.1 (2024-05-13)
 
-fellingdater 0.0.0.9004 (2024-03-21)
-====================================
+-   Paper.md updated for 'The Journal of Open Source Software' (JOSS).
+-   New vignette (article): "The fellingdater workflow"
+
+# fellingdater 1.0.0 (2024-04-16)
+
+-   Major release of `fellingdater` to reflect the fact that the package is now fully functional and has successfully passed the rOpenSci review process.
+-   review history: <https://github.com/ropensci/software-review/issues/618>
+
+# fellingdater 0.0.0.9004 (2024-03-21)
 
 ### REVIEW
-  * new version of the package after review
-  * review history: https://github.com/ropensci/software-review/issues/618
-  * `fellingdateR`has been renamed to `fellingdater`
 
+-   new version of the package after review
+-   review history: <https://github.com/ropensci/software-review/issues/618>
+-   `fellingdateR`has been renamed to `fellingdater`
 
-fellingdater 0.0.0.9003 (2023-11-20)
-====================================
+# fellingdater 0.0.0.9003 (2023-11-20)
 
 ### NEW FEATURES
 
-  * Version ready for submission to rOpenSci for software review
-  * First draft op paper for 'The Journal of Open Source Software' in /paper/paper.md
-  
-  
-
-  
- 
+-   Version ready for submission to rOpenSci for software review
+-   First draft op paper for 'The Journal of Open Source Software' in /paper/paper.md
