+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: fellingdater
 Title: Estimate, report and combine felling dates of historical tree-ring 
     series
-Version: 1.0.1
+Version: 1.0.2
 Authors@R: c(
     person("Kristof", "Haneca", , "Kristof.Haneca@vlaanderen.be",role = c("aut", "cre", "cph"),
         comment = c(ORCID = "0000-0002-7719-8305")),
