### 2024-05-22 #6716 @arfon
@njtierney, @ajpelu – many thanks for your reviews here! JOSS (and rOpenSci) relies upon the volunteer effort of people like you and we simply wouldn't be able to do this without you ✨
